Governance Idealtypes: Preliminary Research
Hi RichardRed0x,

Here is the state of our preliminary research so far on governance idealtypes. 

Considering the complexity of the topic, we thought it would be appropriate to examine the academic and gray literature on this topic to see if we can find a starting point.

We decided that the starting point for a governance framework analysis of a blockchain community should first start with choosing a governance idealtype(s). After a governance idealtype(s) has been chosen, then an appropriate analytic framework should be applied to the blockchain community.

This is our current work so far on the first part, governance idealtypes.

If you have any feedback, please send me a message.